26-09-2023 Eduardo Lopez ADD devolver precio con decimales de guias

+++ DeliveryBackOffice(DEV)/dbo/Stored Procedures/sps_DeliveryOrderDetailIds.sql
@@ -23,7 +23,7 @@
 AS 
 BEGIN
 
-	DECLARE @PriceGuide INT;
+	DECLARE @PriceGuide DECIMAL(14, 2);
 
 	DECLARE @OriginGuideSystem INT = 
 	(
